refactor: move hhtp-helper to a folder
