+++ src/presentation/controllers/login/login.spec.ts
@@ -1,6 +1,6 @@
 import { LoginController } from './login'
 import { MissingParamError } from '../../errors'
-import { badRequest, ok, serverError, unauthorized } from '../../helpers/http-helper'
+import { badRequest, ok, serverError, unauthorized } from '../../helpers/http/http-helper'
 import { HttpRequest, Authentication, Validation } from './login-protocols'
 
 const makeFakeRequest = (): HttpRequest => ({
