+++ src/presentation/controllers/signup/signup.ts
@@ -1,5 +1,5 @@
 import { HttpResponse, HttpRequest, Controller, AddAccount } from './signup-protocols'
-import { badRequest, ok, serverError } from '../../helpers/http-helper'
+import { badRequest, ok, serverError } from '../../helpers/http/http-helper'
 import { Validation } from '../../helpers/validators/validation'
 
 export class SignUpController implements Controller {
